Make distro usable without Dist::Zilla (#11)
While Dist::Zilla is great for CPAN authors, it can make it hard for people to use a distribution's repository unless they have it installed, and installing it can sometimes be quite a hassle.

This commit makes it so that the repository is usable without having or using Dist::Zilla. This means that contributors can just contribute without worrying about authoring tools, and installers like cpanm can install directly from the repo without needing the authoring tool to build the distribution first.